requirements.txt
streamlit
pandas
sqlalchemy
plotly
python-docx
reportlab
